Add files via upload
This excel files consist of inventory details from week 1 to week 13 and week 14 to week 26. From the given data, the inventory amount at the start of the week, end of the week, stock-to-sales and sell through percentage, the anomalies happened in the week 1- week 13 was explained. I was able to forecast financial planning for the upcoming week(15-26) by calculating sales, ideal inventory amount at the start of the week and receipt values which is when to restock items and how much quantity to be restocked. This completes the financial forecasting of week 15-26 by referring to the historical data of the week 1-week 13